qc/modules/cellranger_count: consolidate pipeline tests for 'cellranger_count' mdule.
